test: remove mock tests superseded by testcontainers
